Added Cell-Centered helmsolve
We have a new, broadly general helmholtz solver now.

It solves the problem

(a - b*L)u = f

on a cell-centered grid. It has the ability to do Dirichlet, Neumann, or Periodic boundary conditions (and handles x/y boundary conditions separately).

I also added an upwind evaluator. This assumes that the concentration is on a cell-centered grid, while the velocities are on a staggered grid.

The next step is go go through the actual simulations and modify them to use these grids and call these functions.