Adding different BC to the gradient step
I added new routines which compute the differentiation matrix for homogeneous dirichlet and neumann boundary conditions (in 1D).

The actual gradient step now has switches based on the BCs in each direction, then applies them to rows/columns of the grid function as appropriate.

This all assumes that the data is on a cell-centered grid.

Note: this doesn't really matter for periodic, but matters a fair amount for Neumann and Dirichlet.