Don't need the "gridproblem" construct anymore
Now that helmsolve.m is written, we don't need to pass around the "gridfunction" string anymore. Everything is handled by helmsolve.
However, now the "grid" struct needs to contain fields which are "bcx" and "bcy". These should be strings which are used to activate switches in the various differential/integral routines.